+++ test/unit/helper.js
@@ -8,7 +8,7 @@ function uuid() {
 
 function generateParagraph() {
   const nouns = ["bird", "clock", "boy", "plastic", "duck", "teacher", "lady", "professor", "hamster", "dog"]
-  const verbs = ["kicked", "ran", "flew", "dodged", "sliced", "rolled", "died", "breathed", "slept", "killed"]
+  const verbs = ["kicked", "ran", "flew", "dodged", "sliced", "rolled", "died", "breathed", "slept"]
   const adjectives = ["beautiful", "lazy", "professional", "lovely", "cheerful", "rough", "soft", "hot", "vibrating", "slimy"]
   const adverbs = ["slowly", "elegantly", "precisely", "quickly", "sadly", "humbly", "proudly", "shockingly", "calmly", "passionately"]
   const preposition = ["down", "into", "up", "on", "upon", "below", "above", "through", "across", "towards"]